WIP allay map renderer
The rendering itself happens in AllayMapRenderer but it just delegates
to Vanilla's map renderer - we need to invoke it manually because our
map is not recognized as a map automatically.

Since it is a map, we want all instances of that map to have the same
data (so if you clone the map, it should have the same data). To achieve
this we need to store the data globally, keyed with the mapId. This
storage is implemented via a level capability.

The data we store are just the two targets an allay has to move to,
source and destination. To render them on the map we need to add custom
decorations. As long as a world is active the Vanilla map will hold
those decorations and also automatically sync them to the client, so we
do not have to care about this. Therefor we just need to make sure that
we apply those decorations whenever the targets changed. This happens in
AllayMapItem#inventoryTick.

As already mentioned Vanilla passes the decorations to the client side
automatically, but we still need to sync the targets so that we can show
them the item's tooltip. To do this there are two approaches:
 - sync the capability state to the client
 - sync the item's nbt tag with the AllayMapData state and let Vanilla
   handle the syncing
We currently *try* to use the first approach because it allows side
agnostic data access.

This commit is WIP because somehow the update message we send causes
Minecraft to error out. And because I cannot fix it myself I push this
commit in the hope someone else can fix it :)

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>