Add player-specific map data dirty flags
By storing exactly which player needs syncing of what maps we can
drastically reduce the amount of packets we have to send.

Vanilla does this as well, but they actively track which players
currently hold which maps. This allows them to distribute the updates
only to the interested players.

Our approach instead flags each player in the world but then only
distributes updates if a player is asking for it (by ticking the map
itemstack in their inventory).

I think this needs some adjustments to account for players that are not
in the world at the time of updating -> will be done next.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>