Invalidate sync flags on player world join
The flags can be thought of as "the maps from which we know the player
has accurate and actual data about". As such it makes sense that when
the player joins the world, we do not know if their data is up-to-date
because it could have changed in the meantime.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>